+++ sqlr/types.go
@@ -1,4 +1,4 @@
-package sqlx
+package sqlr
 
 import (
 	"time"
